备份数据库，修改playlist表、创建删除歌单、新增歌曲到歌单

+++ src/main/java/com/example/comelymusic/generate/service/impl/PlaylistServiceImpl.java
@@ -8,11 +8,15 @@ import com.example.comelymusic.generate.controller.requests.PlaylistUpdateReques
 import com.example.comelymusic.generate.entity.Playlist;
 import com.example.comelymusic.generate.entity.PlaylistMusic;
 import com.example.comelymusic.generate.entity.User;
+import com.example.comelymusic.generate.enums.UserPlaylistRelation;
 import com.example.comelymusic.generate.mapper.PlaylistMapper;
 import com.example.comelymusic.generate.mapper.PlaylistMusicMapper;
 import com.example.comelymusic.generate.service.PlaylistService;
+import com.example.comelymusic.generate.service.UserPlaylistService;
 import com.example.comelymusic.generate.service.UserService;
+import com.sun.media.jfxmedia.events.PlayerTimeListener;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 
 /**
@@ -34,21 +38,54 @@ public class PlaylistServiceImpl extends ServiceImpl<PlaylistMapper, Playlist> i
     @Autowired
     private PlaylistMusicMapper playlistMusicMapper;
 
+    @Lazy
+    @Autowired
+    private UserPlaylistService userPlaylistService;
+
     /**
-     * 创建歌单
+     * （创建歌单,并创建用户歌单关系），原子操作，返回结果
      */
     @Override
     public int create(PlaylistCreateRequest request) {
+        if (checkoutDuplicate(request.getName(), request.getUsername())) {
+            // 用户不能创建重复歌单名
+            return -1;
+        }
         Playlist playlist = createRequest2Entity(request);
-        return playlistMapper.insert(playlist);
+        int insertPlaylist = playlistMapper.insert(playlist);
+        if (insertPlaylist == 0) {
+            return 0;
+        }
+        int insertRelation = userPlaylistService.create(request.getName(), request.getUsername(),
+                UserPlaylistRelation.CREATE.getRelation());
+        if (insertRelation == 0) {
+            // 如果关系为创建成功，那删除刚刚创建的歌单，相当于一次原子操作
+            deletePlaylist(new PlaylistSelectRequest().setPlaylistName(request.getName()).setUsername(request.getUsername()));
+            return 0;
+        }
+        return 1;
+    }
+
+    private boolean checkoutDuplicate(String name, String username) {
+        User user = userService.selectByUsername(username);
+        QueryWrapper<Playlist> wrapper = new QueryWrapper<>();
+        wrapper.eq("name", name);
+        wrapper.eq("created_by_user_id", user.getId());
+        Playlist playlist = playlistMapper.selectOne(wrapper);
+        return playlist != null;
     }
 
     /**
-     * 根据创建用户username和歌单名称来删除歌单
+     * 根据创建用户username和歌单名称来删除歌单,并删除用户-歌单关系
      */
     @Override
     public int deletePlaylist(PlaylistSelectRequest request) {
+        // 先删关系表，再删歌单，否则会触发SQL外键异常
+        userPlaylistService.delete(request.getPlaylistName(), request.getUsername());
         Playlist playlist = selectPlaylist(request);
+        if (playlist == null) {
+            return 0;
+        }
         Integer musicNum = playlist.getMusicNum();
         int delete = playlistMapper.deleteById(playlist.getId());
 
@@ -78,8 +115,8 @@ public class PlaylistServiceImpl extends ServiceImpl<PlaylistMapper, Playlist> i
                     playlist.setCreatedByUserId(user.getId());
                 }
             }
-            if (request.getStatus() != null && !request.getStatus().equals(playlist.getStatus())) {
-                playlist.setStatus(request.getStatus());
+            if (request.getVisibility() != null && !request.getVisibility().equals(playlist.getVisibility())) {
+                playlist.setVisibility(request.getVisibility());
             }
             if (request.getDescription() != null) {
                 playlist.setDescription(request.getDescription());
@@ -134,6 +171,7 @@ public class PlaylistServiceImpl extends ServiceImpl<PlaylistMapper, Playlist> i
         return playlistMusicMapper.delete(wrapper);
     }
 
+
     private Playlist createRequest2Entity(PlaylistCreateRequest request) {
         Playlist playlist = new Playlist();
         String name = request.getName();
@@ -152,8 +190,8 @@ public class PlaylistServiceImpl extends ServiceImpl<PlaylistMapper, Playlist> i
         if (request.getDescription() != null) {
             playlist.setDescription(request.getDescription());
         }
-        if (request.getStatus() != null) {
-            playlist.setStatus(request.getStatus());
+        if (request.getVisibility() != null) {
+            playlist.setVisibility(request.getVisibility());
         }
         if (request.getMusicNum() != null) {
             playlist.setMusicNum(request.getMusicNum());
